+++ internal/testutil/testutil.go
@@ -151,6 +151,7 @@ func NewApp(dataDir string) (*pocketbase.PocketBase, error) {
 
 	hooks.RegisterOrgCode(app, orgCollection)
 	hooks.RegisterCodes(app)
+	hooks.RegisterLocationPath(app)
 	hooks.RegisterOrgProvisioning(app, hooks.OrgProvisioningOptions{
 		OrgCollection:                 orgCollection,
 		NatsAccountCollection:         natsOpts.AccountCollectionName,